fix: resolve all production mypy type errors across server/ and client/
- Fix StorageBackend protocol to match concrete implementations: add 'tool'
  role and tool_call_id to create_message; add status/agent_name to
  update_session; add agent_name to MemoryStorageBackend.create_session
- Rename AgentDefinitionRegistry.list() -> get_all() to stop shadowing
  the builtin list type (caused valid-type errors in subagents/primaries)
- Remove dead duplicate kwargs block in llm/registry.py after return
- Add _run() to BrowserTool to satisfy BaseTool abstract method contract
- Fix sandbox_backend.py: type _docker_backend as Any to allow execute()
- Fix mcp_adapter.py: suppress call-overload and no-any-return on create_model
- Fix api/middleware.py: cast call_next result to Response
- Fix api/scoping.py: add return type and **headers: Any annotation
- Fix api/sse.py: annotate data dict to prevent type widening
- Fix storage/migrations.py: pass non-None default to walk_revisions
- Add type: ignore comments for untyped third-party imports (yaml, botocore,
  langchain_aws) and unavoidable Any returns (mlflow, runtime.abort)
- Update ROADMAP.md with bug fixes and mypy improvement entry

All 343 tests pass.

+++ server/app/rate_limiter.py
@@ -52,7 +52,7 @@ class TokenBucket:
             elapsed = now - self.last_update
 
             # Add tokens based on elapsed time
-            self.tokens = min(self.capacity, self.tokens + elapsed * self.rate)
+            self.tokens = int(min(self.capacity, self.tokens + elapsed * self.rate))
             self.last_update = now
 
             if self.tokens >= 1:
